// frontend/src/inc/Api/Status/StatusCodes.ts
/**
 * StatusCodes
 */
export enum StatusCodes {
    OK = '200',
    UNAUTHORIZED = '401',
    FORBIDDEN = '403',
    INTERNAL_ERROR = '500',
    BAD_REQUEST = '400',
    PAYMENT_REQUIRED = '402',
    NOT_FOUND = '404',
    METHOD_NOT_ALLOWED = '405',
    NOT_ACCEPTABLE = '406',
    REQUEST_TIMEOUT = '408',
    CONFLICT = '409',
    GONE = '410'
}